fix (srdf): WIP - different beahviours fro srdf writing
Xbot and Moveit require different srdf. To handle this the state pattern for the control plugin  was used but is not a clean solution. A subclass SRDFWriter could be defined to handle this either ina similar state pattern or by implementing an abstract factory (https://refactoring.guru/design-patterns/abstract-factory/python/example)