fix-wip (gripper URDF): fingers joints should be "fixed" for XBot compatibility
Since the gripper joints are not directly controlled by XBotCore and we do not want these joints to appear in the robot Jacobian we make them fixed.
The fingers would be moved by an external ROS node.
A better way would be to not include the fingers at all (have only their mass in one big body with the gripper). In this way there would not be confusion with the TF.